requirements.txt
PySide6>=6.5.0
pyshark>=0.4.3
scapy>=2.4.5
python-nmap>=0.7.1
mac-vendor-lookup
requests>=2.31.0
psutil>=5.8.0
pyqtgraph>=0.13.0
matplotlib>=3.5.0
PyYAML>=6.0
# ML-IDS dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
seaborn>=0.12.0
# Attack Graph + GeoIP
networkx>=3.0
geoip2>=4.6.0  # optional — only needed if using local GeoLite2-City.mmdb